+++ libraries/rush-lib/src/cli/actions/UpdateAction.ts
@@ -95,7 +95,8 @@ export class UpdateAction extends BaseInstallAction {
       checkOnly: false,
       subspace: this.getTargetSubspace(),
 
-      beforeInstallAsync: () => this.rushSession.hooks.beforeInstall.promise(this)
+      beforeInstallAsync: () => this.rushSession.hooks.beforeInstall.promise(this),
+      terminal: this._terminal
     };
   }
 }