+++ apps/rush/CHANGELOG.md
@@ -1,6 +1,48 @@
 # Change Log - @microsoft/rush
 
-This log was last generated on Wed, 20 Mar 2024 20:45:18 GMT and should not be manually modified.
+This log was last generated on Wed, 27 Mar 2024 05:31:17 GMT and should not be manually modified.
+
+## 5.118.6
+Wed, 27 Mar 2024 05:31:17 GMT
+
+### Updates
+
+- Symlinks are now generated for workspace projects in the temp folder when subspaces and splitWorkspaceCompatibility is enabled.
+
+## 5.118.5
+Tue, 26 Mar 2024 19:58:40 GMT
+
+### Updates
+
+- Use pnpm-sync-lib logging APIs to customize the log message for pnpm-sync operations
+
+## 5.118.4
+Tue, 26 Mar 2024 02:39:06 GMT
+
+### Updates
+
+- Added warnings if there are .npmrc or .pnpmfile.cjs files in project folders after migrating to subspaces
+
+## 5.118.3
+Sat, 23 Mar 2024 01:41:10 GMT
+
+### Updates
+
+- Fix an edge case for computing the PNPM store path when the experimental subspaces feature is enabled
+
+## 5.118.2
+Fri, 22 Mar 2024 17:30:47 GMT
+
+### Updates
+
+- Fix bugs related to path operation in Windows OS for subspace feature
+
+## 5.118.1
+Thu, 21 Mar 2024 16:39:32 GMT
+
+### Updates
+
+- Support PNPM injected installation in Rush subspace feature
 
 ## 5.118.0
 Wed, 20 Mar 2024 20:45:18 GMT